Refactor to minimize Crowd requests
The motivation for this refactor is to add a "crowd_auth_every" option
that declares how often users should be re-authorized with Crowd.  By
default every request is re-authenticated.  Adding crowd
  crowd_auth_every 10.minutes
to your authlogic session class will cause authenticated users to be
re-authenticated every 10 minutes.

I also cleaned up some of the crowd_client stuff so that application
tokens are reused as often as possible (thus reducing Crowd requests).

This temporarily removes support for automatically creating Crowd users
when new local users are added.  This should be added back in a future
commit.